Added backend
It will get a random free port to run express on.

Renderer asks on which port it is running (using ipcRenderer.send("GET_API_ENDPOINT")) and the server will respond to that event with "RESPONSE_API_ENDPOINT" plus
the port value. Finally renderer will listen for "RESPONSE_API_ENDPOINT", and get the port value; thus setting the axios instance baseURL.