Added Welcome page, added zustand, added WelcomeStore, added ServerStore
Removed custom-electron-titlebar (actual garbage).

The flow now is to load the renderer even if the server hasn't started
yet, but using zustand's magic now the renderer can know globally if the
server is loading or loaded, thus, everything will be rendered
instantly, but for example buttons that triggered an action that
required interaction with the backend will now be disabled until the
server state is = loaded.

Once you have visited the Welcome page and logged in or continued
without an account, the key "welcome" will be saved to localStorage,
this will be checked everytime when the app starts, if it's true user
will be taken to home page, otherwise, to the welcome page.